Split crawl into multiple files
