+++ packages/word-definitions/src/crawl/crawl.ts
@@ -1,30 +1,8 @@
 import { Locale } from '@scrabble-solver/types';
 
-import { request } from '../lib';
-
-const crawlFrench = (word: string): Promise<string> => {
-  return request({
-    protocol: 'https',
-    hostname: 'www.cnrtl.fr',
-    path: `/definition/${encodeURIComponent(word)}`,
-  });
-};
-
-const crawlEnglish = (word: string): Promise<string> => {
-  return request({
-    protocol: 'https',
-    hostname: 'www.merriam-webster.com',
-    path: `/dictionary/${encodeURIComponent(word)}`,
-  });
-};
-
-const crawlPolish = (word: string): Promise<string> => {
-  return request({
-    protocol: 'https',
-    hostname: 'sjp.pl',
-    path: `/${encodeURIComponent(word)}`,
-  });
-};
+import crawlEnglish from './crawlEnglish';
+import crawlFrench from './crawlFrench';
+import crawlPolish from './crawlPolish';
 
 const crawlPerLocale: Record<Locale, (word: string) => Promise<string>> = {
   [Locale.EN_GB]: crawlEnglish,
