+++ packages/word-definitions/src/crawl/crawlEnglish.ts
@@ -1,0 +1,11 @@
+import { request } from '../lib';
+
+const crawlEnglish = (word: string): Promise<string> => {
+  return request({
+    protocol: 'https',
+    hostname: 'www.merriam-webster.com',
+    path: `/dictionary/${encodeURIComponent(word)}`,
+  });
+};
+
+export default crawlEnglish;
