+++ packages/word-definitions/src/crawl/crawlFrench.ts
@@ -1,0 +1,11 @@
+import { request } from '../lib';
+
+const crawlFrench = (word: string): Promise<string> => {
+  return request({
+    protocol: 'https',
+    hostname: 'www.cnrtl.fr',
+    path: `/definition/${encodeURIComponent(word)}`,
+  });
+};
+
+export default crawlFrench;
