+++ packages/word-definitions/src/crawl/crawlPolish.ts
@@ -1,0 +1,11 @@
+import { request } from '../lib';
+
+const crawlPolish = (word: string): Promise<string> => {
+  return request({
+    protocol: 'https',
+    hostname: 'sjp.pl',
+    path: `/${encodeURIComponent(word)}`,
+  });
+};
+
+export default crawlPolish;
